Keep data wen resolution changes in correlation plot
Fix a bug in the length of the beam in the bar plot.
